requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
Pillow
python-multipart
pypdf
pytest
pytest-mock
requests
langgraph-checkpoint-sqlite
